extended: refactor deployment tests
